Move CompilationExtensions to Extensions folder
